solving problems of hooks
